Increase ev.clicktime for touchpad usability
When using tap-to-click on a laptop touchpad, there is a delay between
the mousedown and mouseup events. At least on my laptop, this delay is
significant: no matter how quickly I tap, the mouse events are spaced
apart by about 170-180 ms. This is longer than what ev.clicktime allows,
meaning tap-to-click doesn't register as a click to Decker; this makes
things like menus slightly frustrating to operate on a touchpad because
of the difference in behavior compared to clicking a regular mouse.

This commit increases ev.clicktime from 10 to 12 frames; at 60 fps, this
bumps up the click-recognition delay to an even 200 ms.